- Made User ignore functionality realtime. Dont require app restart now
