--- shared/domain/src/commonMain/kotlin/network/bisq/mobile/client/service/user_profile/IgnoreUserProfileRequest.kt
@@ -1,8 +1,0 @@
-package network.bisq.mobile.client.service.user_profile
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class IgnoreUserProfileRequest(
-    val profileId: String,
-)
