+++ shared/domain/src/commonMain/resources/mobile/mobile.properties
@@ -266,5 +266,5 @@ mobile.bisqEasy.takeOffer.paymentMethods.headline.fiat=Which payment method do y
 mobile.bisqEasy.takeOffer.progress.baseSidePaymentMethod=Settlement Method
 mobile.bisqEasy.takeOffer.paymentMethods.headline.btc=Which settlement method do you want to use?
 
-mobile.chat.ignoreUserWarn=Selecting ''Ignore user'' will effectively hide all messages from this user.\n\n\This action will take effect when you restart the app.\n\nTo reverse this, go to Settings > Ignored profiles.
-mobile.chat.undoIgnoreUserWarn=Are yoou sure want to unblock this user?\n\n\This action will take effect when you restart the app.
+mobile.chat.ignoreUserWarn=Selecting ''Ignore user'' will effectively hide all messages from this user.\n\n\To reverse this, go to Settings > Ignored profiles.
+mobile.chat.undoIgnoreUserWarn=Are you sure want to unblock this user?\n\n\This action will take effect when you restart the app.
