+++ shared/domain/src/commonMain/resources/mobile/mobile_es.properties
@@ -243,5 +243,5 @@ mobile.bisqEasy.takeOffer.paymentMethods.headline.fiat=[ES] Which payment method
 mobile.bisqEasy.takeOffer.progress.baseSidePaymentMethod=[ES] Settlement Method
 mobile.bisqEasy.takeOffer.paymentMethods.headline.btc=[ES] Which settlement method do you want to use?
 
-mobile.chat.ignoreUserWarn=[ES] Selecting ''Ignore user'' will effectively hide all messages from this user.\n\n\This action will take effect when you restart the app.\n\nTo reverse this, go to Settings > Ignored profiles.
+mobile.chat.ignoreUserWarn=[ES] Selecting ''Ignore user'' will effectively hide all messages from this user.\n\n\To reverse this, go to Settings > Ignored profiles.
 mobile.chat.undoIgnoreUserWarn=[ES] Are you sure want to unblock this user?\n\n\This action will take effect when you restart the app.
